requirements.txt
torch
torchvision
flask
opencv-python
pillow
numpy